fix: wrap remaining trunc/floor calls in f32() to prevent WGSL i32→f32 warnings
- checksVar.tsx: floor(round(...)) nested pattern with abstractFloat from cs = 1.0/(P.size + 1.0e-6)
- fan2.tsx: trunc((2.0*theta*p2)/p1) with abstractFloat from integer 2.0 * f32 expression
- pie.tsx: trunc(r1*p1 + 0.5) with abstractFloat from + 0.5 literal
- rings2.tsx: trunc((r+p)/twop) where twop = 2*p produces abstractFloat

+++ packages/app/src/flame/variations/parametric/general/checksVar.tsx
@@ -38,7 +38,7 @@ export const checksVar = parametricVariation(
     const cs = 1.0 / (P.size + 1.0e-6)
     const ncx = P.x * -1.0
     const ncy = P.y * -1.0
-    const isXY = floor(round(pos.x * cs)) + floor(round(pos.y * cs))
+    const isXY = f32(floor(round(pos.x * cs))) + f32(floor(round(pos.y * cs)))
     const rnx = P.rnd * random()
     const rny = P.rnd * random()
     const isEven = isXY % 2.0 === 0.0
